# perf(iOSmacOS): Directly convert from `Color` to `CGColor`