--- a/src/Uno.UWP/UI/Color.iOS.cs
+++ b/src/Uno.UWP/UI/Color.iOS.cs
@@ -11,7 +11,7 @@ namespace Windows.UI
 	public partial struct Color : IFormattable
 	{
 		public static implicit operator UIKit.UIColor(Color color) => UIKit.UIColor.FromRGBA(color.R, color.G, color.B, color.A);
-		public static implicit operator CGColor(Color color) => new CGColor(color.R / 255.0f, color.G / 255.0f, color.B / 255.0f, color.A / 255.0f);
+		public static implicit operator CGColor(Color color) => CGColor.CreateSrgb(color.R / 255.0f, color.G / 255.0f, color.B / 255.0f, color.A / 255.0f);
 
 		public static implicit operator Color(UIKit.UIColor color) => color.CGColor;