@MartinZikmund 2022-06-17
@spouliot The conventional commits validation is failing, probably you need to remove the comma and space in commit scope (e.g. `perf(iOSmacOS)`)