@MartinZikmund 2022-06-20
@spouliot seems like iOS UITests are failing - the colors are slightly different than before:

<img width="327" alt="image" src="https://user-images.githubusercontent.com/1075116/174593103-5027ef97-e560-4281-9738-0f76d0167aee.png">