@spouliot 2022-06-21
@MartinZikmund weird, that looks too large for float-vs-nfloat, maybe they don't default to the same colourspace... I'll 👀 asap, thanks!