@spouliot 2022-06-21
Components seems identical on macOS but the colourspace does differ: `kCGColorSpaceSRGB` (from `NSColor`) vs `kCGColorSpaceGenericRGB` (for manual conversion) on the created `CGColor` instances.