@spouliot 2022-06-22
same for iOS, using the same colourspace seems to work fine. Nice test BTW 😄